Initial commit of the native menu framework AWT implementation.  The basic structure is in place and menus are being created properly but a number of issues still need to be addressed:
-- Similar to the Cocoa NMF implementation, the AWT menus are not updated in
real-time by the VCL code to propogate changes to menu item names, enabled/
disabled state, and dynamic menu contents in realtime.  VCL relies on receiving
"menu about to be displayed" events to trigger updates, but AWT menus cannot
generate these events.  A "heartbeat" thread still needs to be added to
continually trigger updates for the frontmost frame.

-- VCLAWTCheckboxMenuItems do not seem to be fully functioning as exceptions
are being thrown in checkItem().  Unsure if this is a result of the above or
not.

-- The LINK between the AWT menu item and the underlying VCL item is non-
functional unless the VCL menu (in-window menu) has been made visible prior to
selecting the item from the native menu.  The addition of the "hearbeat" thread
should remove this synchronization bug.

-- In-window menubars are still visible (e.g. VisibleMenuBar() still returns
FALSE).  This was consciously done to continue to allow for further debugging
and manually triggering menu synchronization between the two environments
until the "heartbeat" thread is in place.

-- The extent of object leakage due to the force regeneration of menubars on
realtime class morphing is unknown.

-- Has not been tested with the most up-to-date Alpha 2 sources and may result
in crashes.

Edward Peterlin
OPENSTEP@neooffice.org