Bugs in the Mac OS X 1.3.1 VM in the CheckboxMenuItem prevent the state of the check in the menu being synchronized properly if setState() is invoked prior to the creation of the underlying AWT peer itself (com.apple.mrj.internal...) or if the state is set to true in the constructor.  We therefore need to delay calls to setState() until after the peer is constructed in order for the state to be correct.
All calls to setState() for CheckboxMenuItems now explicitly check for the
existance of the peer before making the method call.

A new method, syncCheckboxMenuItemState(), was added to the VCLMenuBar.  This
method is invoked from UpdateMenusForFrame().  This placement should ensure
the checkmarks should be correct before the user first sees the menus.

Bug 182

Edward Peterlin
OPENSTEP@neooffice.org

(with lotsa help from Patrick)