As it turns out, there is a possibility that menus may be placed into multiple menubars.  When we're cleaning up a menubar, therefore, we need to destroy/cleanup only the Menu objects associated with that menubar or else the menu peers in other menubars may have their contents inadvertently removed.
A new method, unregisterAWTPeer(), will remove the references to and perform
cleanup for a single AWT Menu/MenuItem object at a time.  The only time
when all of the peers will be removed is when destroy() is invoked and the
associated C++ object is destructed.

Bug 308
Edward Peterlin
OPENSTEP@neooffice.org