Get the About menu item under the Application menu to display the standard NeoJ about dialog.
Under the Apple 1.3.1 VM, the enabling of the About menu item entry is tied
apparently to the installation of an MRJAboutHandler object.  The about
dialog within OOo, however, is displayed from a handler in SfxApplication
instances which, unfortunately, are inaccessible from the vcl module which
is where the Java 1.3.1 MRJ handler needs to be instantiated.

Within the sfx2 module, a new AppleEvent handler is installed that is
handling events of kCoreEventClass/kAEAbout.  This was originally intended
for the finder, but is handled by NeoJ as well.  Sending an about AppleEvent
to the application will invoke this sfx2 handler and subsequently will display
the SfxApplication's About dialog.

In the vcl module, a VCLAboutHandler class was added to bind to the 1.3.1 VM
MRJAboutHandler interface.  In the 1.3.1 VM we can't send AppleEvents from
Java code;  a new custom event class SALEVENT_ABOUT was created and the
Java handler posts this to the VCLEvent queue.  When this custom type is
seen in VCLEventQueueEvent.cxx, a new AppleEvent of kCoreEventClass/kAEAbout
is constructed and sent with "self" as the target.  This queues up the
About invocation so next time through the event loop the sfx2 handler is
invoked and the SfxApplication can be used to respond to the request.

Additional modifications were made in salinst.cxx to allow kAEAbout
AppleEvents to be dispatched using the normal AppleEvent dispatch mechanisms
for Carbon.  No ill effects with the VM Process() crashes were observed on
10.3.7 in limited testing.

Bug 396
Edward Peterlin
OPENSTEP@neooffice.org