Add support for mapping Tools > Options to the traditional Preferences menu item in the NeoOffice/J application menu.  Because VCL locking is required but the handler is in another module, an approach is taken that is similar to the About handler.
An implementation of the MRJPrefsHandler interface is added that posts new
custom SALEVENT_ABOUT events to the VCL event queue.

To handle the SALEEVENT_ABOUT, the VCL module constructs an AppleEvent
of class 'NO%F' and event type 'mPRF'.  This AppleEvent is then sent to
"self", the same process ID.

Next time through the Carbon event loop, the VCL lock is acquired and this
AppleEvent is dispatched.

The hander for this event typ eis installed by the OfficeApplication::Init()
method in the offmgr module.  The handler then invokes the
ExecuteGeneralOptionsDialog() handler with the SID_OPTIONS_TREEDIALOG slot ID,
resulting in the options window being displayed.

Edward Peterlin
OPENSTEP@neooffice.org