Get rid of minimal height restriction in drawPushButton() now that we have the ability to set it in getPreferredPushButtonHeight().
If a pushbutton is requested that is < 30 pixels in width, don't override the
preferred height and let Swing/Aqua LAF draw it as a square bevel button as
most likely it's an icon button.  If the button is > 30 pixels, pass on
Swing's preferred dimensions in getPreferredPushButtonHeight() to force it to
draw as a rounded pushbutton.

Edward Peterlin
OPENSTEP@neooffice.org