Remove the need for strange pointer addition when generating the key for an "as radio group" wrapper by storing such objects in its own separate dictionary with a standard key.
It appears that when getting a wrapper object with create == YES and asRadioGroup == YES, the special pointer addition to the accessible context's key ends up creating a second AquaA11yWrapper instance with the same accessible context in the all wrappers dictionary. This allows the LibreOffice code to create a native radio button group object while sharing the same accessible context with the actual first radio button object.