Make AquaA11yWrapper extend NSAccessibilityElement instead of NSView
Despite various attempts, it appears that NSView instances must be added as a subview of a window or else they are treated as ignored accessible children. Extending NSAccessibilityElement, which is a lightweight class that only extends NSObject, eliminates the huge memory and performance footprint of adding and removing thousands of subviews while using macOS accessibility features.